requirements.txt
playwright>=1.40.0
pydantic>=2.5.0
python-dotenv>=1.0.0
requests>=2.31.0
tenacity>=8.2.0
aiohttp>=3.9.0
asyncio-pool>=0.6.0
tqdm>=4.66.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pyhunter>=1.7
pandas>=2.1.0
openpyxl>=3.1.0
flask>=3.0.0
flask-socketio>=5.3.0
flask-cors>=4.0.0
python-socketio>=5.9.0
eventlet>=0.33.2
